Move the event year from the site title to the masthead title
The duckduckgo search engine grabs the site title from the root page and
displays that title for all of the pages found under the root.

Since the server is setup to redirect the root to the current year, this
means the search results show the current year's title string in the
search results for all of the pages, even for past events.

To fix this, remove the event year from the site title, and put it in
the masthead title instead.  This way the pages show the 2021 year in
the masthead (at the top) but the site title is not leaked on the search
results for previous years.