Makefile: add test and prod targets
Instead of changing the path in the config file, create different
targets for the test and prod paths.  Get the baseurl value directly
from the _config.yml file so we push to the correct year.

Move the test and prod paths to the makefile.
Remove the Makefile.config on make clean.